famtube/requirements.txt
ansicon==1.89.0
arrow==0.17.0
asgiref==3.3.1
blessed==1.17.12
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-common-helpers==0.9.2
django-picklefield==3.0.1
djangorestframework==3.12.2
google-api-core==1.23.0
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jinxed==1.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
rsa==4.6
schedule==0.6.0
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
